requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.8
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
django-crispy-forms==1.13.0
django-filter==21.1
django-phone-field==1.8.1
django-phonenumber-field==6.0.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
docopt==0.6.2
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
packaging==21.3
phonenumberslite==8.12.38
Pillow==8.4.0
pipreqs==0.4.11
pycodestyle==2.8.0
PyJWT==2.3.0
pyparsing==3.0.6
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
yarg==0.1.9
